// internal/az/azreq/grpc/v1/generated/proto/v1/pdp.ts
  export class PolicyStore extends pb_1.Message {
    #one_of_decls: number[][] = [];
    constructor(
      data?:
        | any[]
        | {
            Kind?: string;
            ID?: string;
          }
    ) {
      super();
      pb_1.Message.initialize(
        this,
        Array.isArray(data) ? data : [],
        0,
        -1,
        [],
        this.#one_of_decls
      );
      if (!Array.isArray(data) && typeof data == "object") {
        if ("Kind" in data && data.Kind != undefined) {
          this.Kind = data.Kind;
        }
        if ("ID" in data && data.ID != undefined) {
          this.ID = data.ID;
        }
      }
    }
    get Kind() {
      return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
    }
    set Kind(value: string) {
      pb_1.Message.setField(this, 1, value);
    }
    get ID() {
      return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
    }
    set ID(value: string) {
      pb_1.Message.setField(this, 2, value);
    }
    static fromObject(data: { Kind?: string; ID?: string }): PolicyStore {
      const message = new PolicyStore({});
      if (data.Kind != null) {
        message.Kind = data.Kind;
      }
      if (data.ID != null) {
        message.ID = data.ID;
      }
      return message;
    }
    toObject() {
      const data: {
        Kind?: string;
        ID?: string;
      } = {};
      if (this.Kind != null) {
        data.Kind = this.Kind;
      }
      if (this.ID != null) {
        data.ID = this.ID;
      }
      return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
      const writer = w || new pb_1.BinaryWriter();
      if (this.Kind.length) writer.writeString(1, this.Kind);
      if (this.ID.length) writer.writeString(2, this.ID);
      if (!w) return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): PolicyStore {
      const reader =
          bytes instanceof pb_1.BinaryReader
            ? bytes
            : new pb_1.BinaryReader(bytes),
        message = new PolicyStore();
      while (reader.nextField()) {
        if (reader.isEndGroup()) break;
        switch (reader.getFieldNumber()) {
          case 1:
            message.Kind = reader.readString();
            break;
          case 2:
            message.ID = reader.readString();
            break;
          default:
            reader.skipField();
        }
      }
      return message;
    }
    serializeBinary(): Uint8Array {
      return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): PolicyStore {
      return PolicyStore.deserialize(bytes);
    }
  }
// ...
  export class Entities extends pb_1.Message {
    #one_of_decls: number[][] = [];
    constructor(
      data?:
        | any[]
        | {
            Schema?: string;
            Items?: dependency_1.google.protobuf.Struct[];
          }
    ) {
      super();
      pb_1.Message.initialize(
        this,
        Array.isArray(data) ? data : [],
        0,
        -1,
        [2],
        this.#one_of_decls
      );
      if (!Array.isArray(data) && typeof data == "object") {
        if ("Schema" in data && data.Schema != undefined) {
          this.Schema = data.Schema;
        }
        if ("Items" in data && data.Items != undefined) {
          this.Items = data.Items;
        }
      }
    }
    get Schema() {
      return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
    }
    set Schema(value: string) {
      pb_1.Message.setField(this, 1, value);
    }
    get Items() {
      return pb_1.Message.getRepeatedWrapperField(
        this,
        dependency_1.google.protobuf.Struct,
        2
      ) as dependency_1.google.protobuf.Struct[];
    }
    set Items(value: dependency_1.google.protobuf.Struct[]) {
      pb_1.Message.setRepeatedWrapperField(this, 2, value);
    }
    static fromObject(data: {
      Schema?: string;
      Items?: ReturnType<
        typeof dependency_1.google.protobuf.Struct.prototype.toObject
      >[];
    }): Entities {
      const message = new Entities({});
      if (data.Schema != null) {
        message.Schema = data.Schema;
      }
      if (data.Items != null) {
        message.Items = data.Items.map((item) =>
          dependency_1.google.protobuf.Struct.fromObject(item)
        );
      }
      return message;
    }
    toObject() {
      const data: {
        Schema?: string;
        Items?: ReturnType<
          typeof dependency_1.google.protobuf.Struct.prototype.toObject
        >[];
      } = {};
      if (this.Schema != null) {
        data.Schema = this.Schema;
      }
      if (this.Items != null) {
        data.Items = this.Items.map(
          (item: dependency_1.google.protobuf.Struct) => item.toObject()
        );
      }
      return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
      const writer = w || new pb_1.BinaryWriter();
      if (this.Schema.length) writer.writeString(1, this.Schema);
      if (this.Items.length)
        writer.writeRepeatedMessage(
          2,
          this.Items,
          (item: dependency_1.google.protobuf.Struct) => item.serialize(writer)
        );
      if (!w) return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): Entities {
      const reader =
          bytes instanceof pb_1.BinaryReader
            ? bytes
            : new pb_1.BinaryReader(bytes),
        message = new Entities();
      while (reader.nextField()) {
        if (reader.isEndGroup()) break;
        switch (reader.getFieldNumber()) {
          case 1:
            message.Schema = reader.readString();
            break;
          case 2:
            reader.readMessage(message.Items, () =>
              pb_1.Message.addToRepeatedWrapperField(
                message,
                2,
                dependency_1.google.protobuf.Struct.deserialize(reader),
                dependency_1.google.protobuf.Struct
              )
            );
            break;
          default:
            reader.skipField();
        }
      }
      return message;
    }
    serializeBinary(): Uint8Array {
      return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): Entities {
      return Entities.deserialize(bytes);
    }
  }
// ...
  export class Resource extends pb_1.Message {
    #one_of_decls: number[][] = [[3]];
    constructor(
      data?:
        | any[]
        | ({
            Type?: string;
            ID?: string;
          } & {
            Properties?: dependency_1.google.protobuf.Struct;
          })
    ) {
      super();
      pb_1.Message.initialize(
        this,
        Array.isArray(data) ? data : [],
        0,
        -1,
        [],
        this.#one_of_decls
      );
      if (!Array.isArray(data) && typeof data == "object") {
        if ("Type" in data && data.Type != undefined) {
          this.Type = data.Type;
        }
        if ("ID" in data && data.ID != undefined) {
          this.ID = data.ID;
        }
        if ("Properties" in data && data.Properties != undefined) {
          this.Properties = data.Properties;
        }
      }
    }
    get Type() {
      return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
    }
    set Type(value: string) {
      pb_1.Message.setField(this, 1, value);
    }
    get ID() {
      return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
    }
    set ID(value: string) {
      pb_1.Message.setField(this, 2, value);
    }
    get Properties() {
      return pb_1.Message.getWrapperField(
        this,
        dependency_1.google.protobuf.Struct,
        3
      ) as dependency_1.google.protobuf.Struct;
    }
    set Properties(value: dependency_1.google.protobuf.Struct) {
      pb_1.Message.setOneofWrapperField(this, 3, this.#one_of_decls[0], value);
    }
    get has_Properties() {
      return pb_1.Message.getField(this, 3) != null;
    }
    get _Properties() {
      const cases: {
        [index: number]: "none" | "Properties";
      } = {
        0: "none",
        3: "Properties",
      };
      return cases[pb_1.Message.computeOneofCase(this, [3])];
    }
    static fromObject(data: {
      Type?: string;
      ID?: string;
      Properties?: ReturnType<
        typeof dependency_1.google.protobuf.Struct.prototype.toObject
      >;
    }): Resource {
      const message = new Resource({});
      if (data.Type != null) {
        message.Type = data.Type;
      }
      if (data.ID != null) {
        message.ID = data.ID;
      }
      if (data.Properties != null) {
        message.Properties = dependency_1.google.protobuf.Struct.fromObject(
          data.Properties
        );
      }
      return message;
    }
    toObject() {
      const data: {
        Type?: string;
        ID?: string;
        Properties?: ReturnType<
          typeof dependency_1.google.protobuf.Struct.prototype.toObject
        >;
      } = {};
      if (this.Type != null) {
        data.Type = this.Type;
      }
      if (this.ID != null) {
        data.ID = this.ID;
      }
      if (this.Properties != null) {
        data.Properties = this.Properties.toObject();
      }
      return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
      const writer = w || new pb_1.BinaryWriter();
      if (this.Type.length) writer.writeString(1, this.Type);
      if (this.ID.length) writer.writeString(2, this.ID);
      if (this.has_Properties)
        writer.writeMessage(3, this.Properties, () =>
          this.Properties.serialize(writer)
        );
      if (!w) return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): Resource {
      const reader =
          bytes instanceof pb_1.BinaryReader
            ? bytes
            : new pb_1.BinaryReader(bytes),
        message = new Resource();
      while (reader.nextField()) {
        if (reader.isEndGroup()) break;
        switch (reader.getFieldNumber()) {
          case 1:
            message.Type = reader.readString();
            break;
          case 2:
            message.ID = reader.readString();
            break;
          case 3:
            reader.readMessage(
              message.Properties,
              () =>
                (message.Properties =
                  dependency_1.google.protobuf.Struct.deserialize(reader))
            );
            break;
          default:
            reader.skipField();
        }
      }
      return message;
    }
    serializeBinary(): Uint8Array {
      return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): Resource {
      return Resource.deserialize(bytes);
    }
  }